// examples/demo-eos-seed-app/src/generated/schema.ts
export interface BotCommand {
    _: 'botCommand'
    command: string
    description: string
}

export interface BotInfo {
    _: 'botInfo'
    user_id: number
    description: string
    commands: BotCommand[]
}
// ...
export interface ContactsLink {
    _: 'contacts.link'
    my_link: ContactLink
    foreign_link: ContactLink
    user: User
}
// ...
export interface InputUserEmpty {
    _: 'inputUserEmpty'
}

export interface InputUserSelf {
    _: 'inputUserSelf'
}

export interface InputUserSupport {
    _: 'inputUserSupport'
}

export interface InputUser_ {
    _: 'inputUser'
    user_id: number
    access_hash: bigint
}
// ...
export interface PeerNotifySettings {
    _: 'peerNotifySettings'
    show_previews?: boolean
    silent?: boolean
    mute_until?: number
    sound?: string
}
// ...
export interface UserEmpty {
    _: 'userEmpty'
    id: number
}

export interface UserFull {
    _: 'userFull'
    blocked?: true
    phone_calls_available?: true
    phone_calls_private?: true
    can_pin_message?: true
    user: User
    about?: string
    link: ContactsLink
    profile_photo?: unknown
    notify_settings: PeerNotifySettings
    bot_info?: BotInfo
    pinned_msg_id?: number
    common_chats_count: number
    folder_id?: number
}

export interface UserProfilePhotoEmpty {
    _: 'userProfilePhotoEmpty'
    color_id?: string
}

export interface UserProfilePhoto_ {
    _: 'userProfilePhoto'
    photo_id: bigint
    photo_small: unknown
    photo_big: unknown
    dc_id: number
    url: string
}

export interface UserService {
    _: 'userService'
    id: number
    self?: true
    contact?: true
    type: ServiceAccountType
    first_name?: string
    last_name?: string
    username?: string
    email: string
    eos_name: string
    photo: UserProfilePhoto
    access_hash: bigint
}
// ...
export interface UserUnreged {
    _: 'userUnreged'
    id: number
    contact?: true
    first_name?: string
    last_name?: string
    phone?: string
    access_hash: bigint
}

export interface User_ {
    _: 'user'
    id: number
    self?: true
    contact?: true
    mutual_contact?: true
    blocked?: true
    verified?: true
    buisness?: true
    first_name?: string
    last_name?: string
    username?: string
    phone?: string
    email: string
    eos_name: string
    photo: UserProfilePhoto
    subscriptions?: SubscriptionsCommonInfo
    lang_code?: string
    access_hash: bigint
    kyc_verifed: number
    has_unionpay_account?: true
    status?: UserStatus
}
// ...
export type ContactLink = ContactLinkUnknown | ContactLinkNone | ContactLinkHasPhone | ContactLinkContact
// ...
export type InputUser = InputUserEmpty | InputUserSelf | InputUser_ | InputUserSupport
// ...
export type User = UserEmpty | UserService | UserUnreged | User_
// ...
export interface UsersGetFullUserParams {
    id: InputUser
}